+++ packages/client/components/TeamPrompt/TeamPromptOptionsMenu.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled'
-import {Flag, Replay} from '@mui/icons-material'
+import {Flag, Link, Replay} from '@mui/icons-material'
 import graphql from 'babel-plugin-relay/macro'
 import React from 'react'
 import {useFragment} from 'react-relay'
@@ -8,13 +8,18 @@ import {MenuProps} from '~/hooks/useMenu'
 import useMutationProps from '~/hooks/useMutationProps'
 import useRouter from '~/hooks/useRouter'
 import EndTeamPromptMutation from '~/mutations/EndTeamPromptMutation'
-import StartRecurrenceMutation from '~/mutations/StartRecurrenceMutation'
-import StopRecurrenceMutation from '~/mutations/StopRecurrenceMutation'
 import {TeamPromptOptionsMenu_meeting$key} from '~/__generated__/TeamPromptOptionsMenu_meeting.graphql'
+import SendClientSegmentEventMutation from '../../mutations/SendClientSegmentEventMutation'
 import {PALETTE} from '../../styles/paletteV3'
+import makeAppURL from '../../utils/makeAppURL'
 import Menu from '../Menu'
 import MenuItem from '../MenuItem'
 import {MenuItemLabelStyle} from '../MenuItemLabel'
+
+const LinkIcon = styled(Link)({
+  color: PALETTE.SLATE_600,
+  marginRight: 8
+})
 
 const ReplayIcon = styled(Replay)({
   color: PALETTE.SLATE_600,
@@ -28,23 +33,29 @@ const FlagIcon = styled(Flag)({
 
 const OptionMenuItem = styled('div')({
   ...MenuItemLabelStyle,
-  width: '200px'
+  width: '240px'
 })
 
 interface Props {
   meetingRef: TeamPromptOptionsMenu_meeting$key
   menuProps: MenuProps
+  openRecurrenceSettingsModal: () => void
+  popTooltip: () => void
 }
 
 const TeamPromptOptionsMenu = (props: Props) => {
-  const {meetingRef, menuProps} = props
+  const {meetingRef, menuProps, openRecurrenceSettingsModal, popTooltip} = props
 
   const meeting = useFragment(
     graphql`
       fragment TeamPromptOptionsMenu_meeting on TeamPromptMeeting {
         id
+        team {
+          id
+        }
         meetingSeries {
           id
+          recurrenceRule
           cancelledAt
           activeMeetings {
             id
@@ -56,10 +67,11 @@ const TeamPromptOptionsMenu = (props: Props) => {
     meetingRef
   )
 
-  const {id: meetingId, meetingSeries, endedAt} = meeting
+  const {id: meetingId, meetingSeries, endedAt, team} = meeting
   const atmosphere = useAtmosphere()
   const {onCompleted, onError} = useMutationProps()
   const {history} = useRouter()
+
   const isEnded = !!endedAt
   const hasRecurrenceEnabled = meetingSeries && !meetingSeries.cancelledAt
   const hasActiveMeetings = !!meetingSeries?.activeMeetings?.length
@@ -71,27 +83,44 @@ const TeamPromptOptionsMenu = (props: Props) => {
 
   return (
     <Menu ariaLabel={'Edit the meeting'} {...menuProps}>
+      {hasRecurrenceEnabled && (
+        <MenuItem
+          key='link'
+          label={
+            <OptionMenuItem>
+              <LinkIcon />
+              Copy meeting permalink
+            </OptionMenuItem>
+          }
+          onClick={async () => {
+            popTooltip()
+            menuProps.closePortal()
+            const copyUrl = makeAppURL(window.location.origin, `meeting-series/${meetingId}`)
+            await navigator.clipboard.writeText(copyUrl)
+
+            SendClientSegmentEventMutation(atmosphere, 'Copied Meeting Series Link', {
+              teamId: team?.id,
+              meetingId: meetingId
+            })
+          }}
+        />
+      )}
       <MenuItem
         key='copy'
         isDisabled={!canToggleRecurrence}
         label={
           <OptionMenuItem>
             <ReplayIcon />
-            {hasRecurrenceEnabled ? <span>{'Stop repeating'}</span> : <span>{'Repeat M-F'}</span>}
+            {hasRecurrenceEnabled ? (
+              <span>{'Edit recurrence settings'}</span>
+            ) : (
+              <span>{'Start recurrence'}</span>
+            )}
           </OptionMenuItem>
         }
         onClick={() => {
           menuProps.closePortal()
-
-          if (hasRecurrenceEnabled) {
-            StopRecurrenceMutation(
-              atmosphere,
-              {meetingSeriesId: meetingSeries.id},
-              {onCompleted, onError}
-            )
-          } else {
-            StartRecurrenceMutation(atmosphere, {meetingId}, {onCompleted, onError})
-          }
+          openRecurrenceSettingsModal()
         }}
       />
       <MenuItem