+++ packages/client/subscriptions/NotificationSubscription.ts
@@ -26,6 +26,7 @@ import {
   removeOrgUserNotificationUpdater
 } from '../mutations/RemoveOrgUserMutation'
 import {popNotificationToastOnNext} from '../mutations/toasts/popNotificationToast'
+import {updateNotificationToastOnNext} from '../mutations/toasts/updateNotificationToast'
 import {LocalStorageKey} from '../types/constEnums'
 import {OnNextHandler, OnNextHistoryContext, SharedUpdater} from '../types/relayMutations'
 import {
@@ -87,7 +88,15 @@ const subscription = graphql`
           ...NotificationPicker_notification @relay(mask: false)
         }
       }
+
+      ... on UpdatedNotification {
+        updatedNotification {
+          ...NotificationPicker_notification @relay(mask: false)
+        }
+      }
+
       ...popNotificationToast_notification @relay(mask: false)
+      ...updateNotificationToast_notification @relay(mask: false)
 
       ... on AuthTokenPayload {
         id
@@ -113,7 +122,7 @@ const subscription = graphql`
       }
 
       # Feature flags
-      ... on AddFeatureFlagPayload {
+      ... on UpdateFeatureFlagPayload {
         user {
           id
           # add flag here
@@ -235,7 +244,8 @@ const onNextHandlers = {
   StripeFailPaymentPayload: stripeFailPaymentNotificationOnNext,
   MeetingStageTimeLimitPayload: meetingStageTimeLimitOnNext,
   InvalidateSessionsPayload: invalidateSessionsNotificationOnNext,
-  AddedNotification: popNotificationToastOnNext
+  AddedNotification: popNotificationToastOnNext,
+  UpdatedNotification: updateNotificationToastOnNext
 } as const
 
 const NotificationSubscription = (
@@ -255,7 +265,7 @@ const NotificationSubscription = (
         case 'AcceptTeamInvitationPayload':
           acceptTeamInvitationNotificationUpdater(payload, context)
           break
-        case 'AddFeatureFlagPayload':
+        case 'UpdateFeatureFlagPayload':
           break
         case 'AddNewFeaturePayload':
           addNewFeatureNotificationUpdater(payload, context)