+++ packages/server/database/types/ScheduledJob.ts
@@ -1,6 +1,9 @@
 import generateUID from '../../generateUID'
 
-export type ScheduledJobType = 'MEETING_STAGE_TIME_LIMIT_END' | 'LOCK_ORGANIZATION'
+export type ScheduledJobType =
+  | 'MEETING_STAGE_TIME_LIMIT_END'
+  | 'LOCK_ORGANIZATION'
+  | 'WARN_ORGANIZATION'
 
 export default abstract class ScheduledJob {
   id = generateUID()