+++ packages/server/graphql/public/types/UserFeatureFlags.ts
@@ -6,7 +6,9 @@ const UserFeatureFlags: UserFeatureFlagsResolvers = {
   insights: ({insights}) => !!insights,
   templateLimit: ({templateLimit}) => !!templateLimit,
   aiSummary: ({aiSummary}) => !!aiSummary,
-  meetingHistoryLimit: ({meetingHistoryLimit}) => !!meetingHistoryLimit
+  noMeetingHistoryLimit: ({noMeetingHistoryLimit}) => !!noMeetingHistoryLimit,
+  checkoutFlow: ({checkoutFlow}) => !!checkoutFlow,
+  retrosInDisguise: ({retrosInDisguise}) => !!retrosInDisguise
 }
 
 export default UserFeatureFlags