+++ packages/server/socketHandlers/handleMessage.ts
@@ -1,11 +1,12 @@
 import {OutgoingMessage} from '@mattkrick/graphql-trebuchet-client'
-import {WebSocket} from 'uWebSockets.js'
+import {WebSocketBehavior} from 'uWebSockets.js'
 import handleGraphQLTrebuchetRequest from '../graphql/handleGraphQLTrebuchetRequest'
 import ConnectionContext from '../socketHelpers/ConnectionContext'
 import keepAlive from '../socketHelpers/keepAlive'
 import sendGQLMessage from '../socketHelpers/sendGQLMessage'
 import handleReliableMessage from '../utils/handleReliableMessage'
 import sendToSentry from '../utils/sendToSentry'
+import {SocketUserData} from './handleOpen'
 
 const handleParsedMessage = async (
   parsedMessage: OutgoingMessage,
@@ -25,11 +26,8 @@ const handleParsedMessage = async (
 const PONG = 65
 const isPong = (messageBuffer: Buffer) => messageBuffer.length === 1 && messageBuffer[0] === PONG
 
-const handleMessage = (
-  websocket: WebSocket | {connectionContext: ConnectionContext},
-  message: ArrayBuffer
-) => {
-  const {connectionContext} = websocket
+const handleMessage: WebSocketBehavior<SocketUserData>['message'] = (websocket, message) => {
+  const {connectionContext} = websocket.getUserData()
   const messageBuffer = Buffer.from(message)
   if (isPong(messageBuffer)) {
     keepAlive(connectionContext)
@@ -56,9 +54,13 @@ const handleMessage = (
   }
 }
 
-const safeHandleMessage = (websocket: WebSocket, message: ArrayBuffer) => {
+const safeHandleMessage: WebSocketBehavior<SocketUserData>['message'] = (
+  websocket,
+  message,
+  isBinary
+) => {
   try {
-    handleMessage(websocket, message)
+    handleMessage(websocket, message, isBinary)
   } catch (e) {
     const error = e instanceof Error ? e : new Error('handleMessage failed')
     sendToSentry(error)